engine turn logic

+++ include/board.h
@@ -64,8 +64,8 @@ public:
     using resource_type = resource::Type;
 
 private:
-    inline std::pair<float, float> get_coin_drawing_pos(int x_pos,
-        int y_pos) const noexcept;
+    inline std::pair<float, float> get_coin_drawing_pos(float x_pos,
+        float y_pos) const noexcept;
     bool add_coin(NVGcontext* ctx, Engine::Column col, Engine::Color color);
     void draw_coins(NVGcontext* ctx) const;
